requirements.txt
colorama>=0.4.6
colorlog>=6.10.1
pygame-ce>=2.5.7
